+++ src/rect.rs
@@ -1,12 +1,12 @@
 pub struct Rect {
-    pub x: u32,
-    pub y: u32,
+    pub x: f32,
+    pub y: f32,
     pub width: u32,
     pub height: u32,
 }
 
 impl Rect {
-    pub fn new(x: u32, y: u32, width: u32, height: u32) -> Self {
+    pub fn new(x: f32, y: f32, width: u32, height: u32) -> Self {
         Self {
             x,
             y,
